+++ src/components/atoms/tab/Tab.tsx
@@ -1,44 +1,15 @@
-import React, { useEffect, useState } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import React from 'react';
 
 import type { TabType } from '@/types';
+import useTab from './hooks/useTab';
 import * as S from './Tab.styled';
-
-const QUERY_STRING_TAB_KEY = 'tabId';
 
 interface TabProps {
   tabs: TabType[];
 }
 
 const Tab = ({ tabs }: TabProps) => {
-  const [searchParams, setSearchParams] = useSearchParams();
-
-  const [activeTab, setActiveTab] = useState<TabType | null>(null);
-
-  const handleTabClick =
-    ({ key }: TabType) =>
-    () => {
-      setSearchParams({ [QUERY_STRING_TAB_KEY]: key });
-    };
-
-  useEffect(() => {
-    const queryStringTabValue = searchParams.get(QUERY_STRING_TAB_KEY);
-
-    if (!queryStringTabValue) {
-      setActiveTab(tabs[0]);
-      return;
-    }
-
-    const selectedTab = tabs.filter((tab) => tab.key === queryStringTabValue);
-
-    if (!selectedTab.length) {
-      setActiveTab(tabs[0]);
-      setSearchParams({ [QUERY_STRING_TAB_KEY]: tabs[0].key });
-      return;
-    }
-
-    setActiveTab(selectedTab[0]);
-  }, [searchParams.get(QUERY_STRING_TAB_KEY)]);
+  const { activeTab, handleTabClick } = useTab(tabs);
 
   return (
     <>
